Please turn on GlobalRegistrarAddr (registrar)
I have locally tested the contracted deployed to this address and the registrar functionality. It was the exact bytecode from the original in the variable GlobalRegistrarCode. Original sender of this contract is given no special permissions that I can tell. So, submitting this PR only speeds the process hopefully. No hard feelings if this gets rejected, I'd just really like to see it working. :-)